Did a major revision.

+++ modules/behance.php
@@ -1,9 +1,14 @@
 <?php
 include_once 'module_common.php';
 flush();
-$count = 3;
-$projects_url = "http://www.behance.net/v2/users/tpryan/projects?api_key=lVyYO1uqnBR27R8IkVJkwvDOWKz9qvSm";
-$wips_url = "http://www.behance.net/v2/users/tpryan/wips?api_key=lVyYO1uqnBR27R8IkVJkwvDOWKz9qvSm";
+$count = 1;
+
+$behance = array(
+	    'base' => "http://www.behance.net/v2/",
+	    'user' => "tpryan",
+	    'key' => "lVyYO1uqnBR27R8IkVJkwvDOWKz9qvSm"
+	);
+
 $cache_dir = $_SERVER['DOCUMENT_ROOT'] . "/assets/cache/";
 $service_cache = $cache_dir . "behance.html";
 $cache_age = 2 * 60 * 60;
@@ -20,7 +25,7 @@ if (shouldStillBeCached($service_cache, $cache_age)){
 }
 else{
 	try {
-		$service_html = refreshServiceHTML($projects_url, $wips_url,  $count, $service_cache, $cache_dir);
+		$service_html = refreshServiceHTML($behance,  $count, $service_cache, $cache_dir);
 	} catch (Exception $e) {
 		$service_html = "<article><p>No current projects</p></article>";
 		var_dump($e);
@@ -32,8 +37,8 @@ echo $service_html;
 
 
 
-function refreshServiceHTML($project_url, $wips_url, $count, $service_cache, $cache_dir){
-	$service_content = get_combined_behance_content($project_url, $wips_url);
+function refreshServiceHTML($behance, $count, $service_cache, $cache_dir){
+	$service_content = get_wips($behance);
 	$service_html = generateBehanceHTML($service_content, $count);
 	$cache_html = "<!-- From Cache -->" . $service_html;
 	file_put_contents($service_cache, $cache_html);
@@ -51,45 +56,46 @@ function cache_images($service_content, $cache_dir, $count){
 		$target = $service_content[$i]['img'];
 		$destination = $cache_dir .$service_content[$i]['id']  .   '.jpg';
 
-
+		if (!file_exists($destination)){
+			$image = file_get_contents($target);
+			file_put_contents($destination, $image);
+		}
 		
 
-		$image = file_get_contents($target);
-		file_put_contents($destination, $image);
+		
 		
 	}
 	
 }	
 
 
-function get_combined_behance_content($projects_url,$wips_url){
-	//$projects = get_content_from_service($projects_url);
+function get_wips($behance){
+	$wips_url = $behance['base'] . "users/" . $behance['user'] . "/wips?api_key=" . $behance['key'];
 	$wips = get_content_from_service($wips_url);
 
-	// $projects = json_decode(url_get_contents("http://new.terrenceryan.dev/modules/projects.js"), true);
-	// $wips = json_decode(url_get_contents("http://new.terrenceryan.dev/modules/wips.js"), true);
 
 	$result_array = array();
 
 
-	// for ($i=0; $i<count($projects); $i++){
-	// 	$entry = array();
-	// 	$entry['id']  = "project" . $projects['projects'][$i]['id'];
-	// 	$entry['name']  = $projects['projects'][$i]['name'];
-	// 	$entry['date']  = $projects['projects'][$i]['created_on'];
-	// 	$entry['url']  = $projects['projects'][$i]['url'];
-	// 	$entry['img']  =  $projects['projects'][$i]['covers']['404'];
-	// 	array_push($result_array, $entry);
-	// }
 
 	for ($i=0; $i<count($wips); $i++){
 		$revision = array_shift($wips['wips'][$i]['revisions']);
+		$comment_url = $behance['base'] . "wips/" . $wips['wips'][$i]['id'] . "/" .  $revision['id'] ."/comments?api_key=" . $behance['key'];
+
+		$comments = get_content_from_service($comment_url);
+
+
+		$first_comment = array_pop($comments["comments"]);
 		$entry = array();
 		$entry['id']  = "wip" . $revision['id'];
 		$entry['name']  = $wips['wips'][$i]['title'];
 		$entry['date']  = $wips['wips'][$i]['created_on'];
 		$entry['url']  = $revision['url'];
 		$entry['img']  = $revision['images']['thumbnail']['url'];
+		$entry['desc']  = $first_comment['comment'];
+
+		
+
 		array_push($result_array, $entry);
 	}
 
@@ -128,8 +134,9 @@ function generateBehanceHTML($service_json, $count ){
 		$item .= $date_string;
 		$item .= '</time>' . "\n";
 		$item .= '				<a href="' . $service_json[$i]['url'] . '">'. "\n";
-		$item .= '					<img src="/assets/cache/' . $service_json[$i]['id']  .   '.jpg" width="100%" />'. "\n";
+		$item .= '					<img src="/assets/cache/' . $service_json[$i]['id']  .   '.jpg" />'. "\n";
 		$item .= '				</a>'. "\n";
+		$item .= '				<div><p>' . $service_json[$i]['desc'] . '</p></div>'. "\n";
 		$item .= '			</article>' . "\n";
 		$result .= $item;
 	}
